+++ src/Plainion.Windows/Controls/Text/AutoCorrection/ArrowCorrection.cs
@@ -1,0 +1,61 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Windows.Documents;
+
+namespace Plainion.Windows.Controls.Text.AutoCorrection
+{
+    public class UnicodeSymbolCorrection : IAutoCorrection
+    {
+        public class Symbol
+        {
+            public Symbol(string ascii, string uniCode)
+            {
+                Ascii = ascii;
+                UniCode = uniCode;
+            }
+
+            public string Ascii { get; private set; }
+            public string UniCode { get; private set; }
+        }
+
+        public UnicodeSymbolCorrection()
+        {
+            Symbols = new List<Symbol>();
+            Symbols.Add(new Symbol("-->", "\u2192"));
+            Symbols.Add(new Symbol("==>", "\u21e8"));
+        }
+
+        public IList<Symbol> Symbols { get; private set; }
+
+        public bool TryApply(TextRange range)
+        {
+            bool ret = false;
+
+            foreach (var wordRange in DocumentOperations.GetWords(range))
+            {
+                var symbol = Symbols.FirstOrDefault(x => x.Ascii == wordRange.Text);
+                if (symbol != null)
+                {
+                    wordRange.Text = symbol.UniCode;
+
+                    ret = true;
+                }
+            }
+
+            return ret;
+        }
+
+        public bool TryUndo(TextPointer pos)
+        {
+            var wordRange = DocumentOperations.GetWordAt(pos);
+            var symbol = Symbols.FirstOrDefault(x => x.UniCode == wordRange.Text);
+            if (symbol != null)
+            {
+                wordRange.Text = symbol.Ascii;
+                return true;
+            }
+
+            return false;
+        }
+    }
+}