+++ src/Plainion.Windows/Controls/Text/AutoCorrection/IAutoCorrection.cs
@@ -5,6 +5,6 @@ namespace Plainion.Windows.Controls.Text.AutoCorrection
     public interface IAutoCorrection
     {
         bool TryApply(TextRange range);
-        TextPointer TryUndo(TextPointer start);
+        bool TryUndo(TextPointer pos);
     }
 }